refactor: modularize symbol list feature
