+++ internal/feature/symbollist/domain/entity/symbol.go
