+++ internal/feature/symbollist/transport/http/dto/symbol_list.go
